Added initial support for using readable streams as method return values. Proxies still treat this as a buffer response.
The WebSocket transport doesn't (yet) support streams. Only available in HttpTransport.
Also, the stream type can only be used as a method return type: using it as the type of a method parameter or as part of a structure definition will throw an error.

Bumped version to 2.2.0.

Start testing on node 6.9 as well.