.Net: improved support for handling double struct members. Bumped version to 2.3.8.
A curious bug occurs when a request which contains double struct members
is sent to the server. Zero values are correctly serialized to 0.0, so
a struct would look like { field: 0.0 }.
The RpcTunnel.cs correctly POSTs this to the server as a byte array with
the correct content length. However, the data is changed to { field: 0 }
by the .Net network stack for unknown reasons. This discrepancy was discovered
by using WireShark to inspect the TCP packets in flight.
This, in turn, results in fewer than the specified number of bytes to be
sent to the server. In the end, the client request times out as the
server sits waiting for the remaining few bytes.

The workaround is to serialize all double values as strings, e.g.
{ field: "0.0" }. The server will correctly recognize and parse this value
to a floating-point number in JavaScript.
Using this workaround does not trigger the bug in the .Net network stack.